refactor: Remove unrelated changes to keep PR focused on canvas tools.
