+++ src/tools/get-users.ts
@@ -1,8 +1,8 @@
 import z from "zod";
-import { DEFAULT_LIMIT } from "../structure/constants";
-import MCPUserError from "../structure/MCPUserError";
-import registerTool from "../structure/registerTool";
-import paginationData from "../utils/paginationData";
+import { DEFAULT_LIMIT } from "./structure/constants";
+import MCPUserError from "./structure/MCPUserError";
+import registerTool from "./structure/registerTool";
+import paginationData from "./utils/paginationData";
 
 const description = `
 Return all Users that the user can see in the workspace.
