+++ src/tools/index.ts
@@ -22,6 +22,7 @@ import "./get-provider-resource";
 import "./get-report-notification";
 import "./get-team";
 import "./get-teams";
+import "./get-users";
 import "./list-anomalies";
 import "./list-audit-logs";
 import "./list-cost-integrations";
@@ -42,5 +43,4 @@ import "./submit-user-feedback";
 import "./update-anomaly";
 import "./update-folder";
 import "./update-report-notification";
-import "./users";
 import "./workspaces";
