+++ src/tools/list-costs.test.ts
@@ -1,5 +1,4 @@
-import type { GetCostReportResponse, GetCostsResponse } from "@vantage-sh/vantage-client";
-import { pathEncode } from "@vantage-sh/vantage-client";
+import type { GetCostsResponse } from "@vantage-sh/vantage-client";
 import { expect } from "vitest";
 import tool from "./list-costs";
 import { DEFAULT_LIMIT } from "./structure/constants";
@@ -20,58 +19,16 @@ type OutputSchema = ExtractOutputSchema<typeof tool>;
 
 const DEFAULT_GROUPINGS = "provider,service,account_id" as unknown as string[];
 
-const REPORT_SETTINGS = {
-  include_credits: false,
-  include_refunds: false,
-  include_discounts: true,
-  include_tax: true,
-  amortize: true,
-  unallocated: false,
-  aggregate_by: "cost" as const,
-  show_previous_period: true,
-};
-
-const REPORT_SETTINGS_AS_API_PARAMS = {
-  "settings[aggregate_by]": "cost",
-  "settings[amortize]": true,
-  "settings[include_credits]": false,
-  "settings[include_discounts]": true,
-  "settings[include_refunds]": false,
-  "settings[include_tax]": true,
-  "settings[show_previous_period]": true,
-  "settings[unallocated]": false,
-};
-
-const reportData: GetCostReportResponse = {
-  token: "crt_123",
-  title: "Test Report",
-  settings: REPORT_SETTINGS,
-  groupings: "service,provider",
-  created_at: "2024-07-03T00:00:00Z",
-  workspace_token: "wrkspc_123",
+// Non-settings fields expected in every /v2/costs request.
+const baseApiParams = {
+  page: 1,
+  cost_report_token: "crt_123",
   start_date: "2023-01-01",
   end_date: "2023-01-31",
-  date_interval: "custom",
-  chart_type: "line",
-  date_bin: "cumulative",
-  business_metric_tokens_with_metadata: [],
-  default_forecast: { kind: "baseline" },
-  filter: null,
-  chart_settings: {
-    y_axis_dimension: "cost",
-    x_axis_dimension: ["date"],
-  },
+  date_bin: "month" as const,
+  groupings: DEFAULT_GROUPINGS,
+  limit: DEFAULT_LIMIT,
 };
-
-const getReportCall = (token: string) => ({
-  endpoint: `/v2/cost_reports/${pathEncode(token)}` as const,
-  params: {},
-  method: "GET" as const,
-  result: {
-    ok: true as const,
-    data: reportData,
-  },
-});
 
 const validArguments: InferValidators<Validators> = {
   page: 1,
@@ -151,21 +108,11 @@ const successData: GetCostsResponse = {
 
 const executionTests: ExecutionTestTableItem<Validators, OutputSchema>[] = [
   {
-    name: "successful call with month date_bin uses report settings as defaults",
+    name: "successful call with month date_bin omits settings so API uses report defaults",
     apiCallHandler: requestsInOrder([
-      getReportCall("crt_123"),
       {
         endpoint: "/v2/costs",
-        params: {
-          ...REPORT_SETTINGS_AS_API_PARAMS,
-          cost_report_token: "crt_123",
-          start_date: "2023-01-01",
-          end_date: "2023-01-31",
-          date_bin: "month",
-          groupings: DEFAULT_GROUPINGS,
-          page: 1,
-          limit: DEFAULT_LIMIT,
-        },
+        params: baseApiParams,
         method: "GET",
         result: {
           ok: true,
@@ -190,15 +137,11 @@ const executionTests: ExecutionTestTableItem<Validators, OutputSchema>[] = [
   {
     name: "successful call with day date_bin",
     apiCallHandler: requestsInOrder([
-      getReportCall("crt_123"),
       {
         endpoint: "/v2/costs",
         params: {
-          ...validArguments,
-          ...REPORT_SETTINGS_AS_API_PARAMS,
-          groupings: DEFAULT_GROUPINGS,
+          ...baseApiParams,
           date_bin: "day",
-          limit: DEFAULT_LIMIT,
         },
         method: "GET",
         result: {
@@ -218,15 +161,11 @@ const executionTests: ExecutionTestTableItem<Validators, OutputSchema>[] = [
   {
     name: "successful call with week date_bin",
     apiCallHandler: requestsInOrder([
-      getReportCall("crt_123"),
       {
         endpoint: "/v2/costs",
         params: {
-          ...validArguments,
-          ...REPORT_SETTINGS_AS_API_PARAMS,
-          groupings: DEFAULT_GROUPINGS,
+          ...baseApiParams,
           date_bin: "week",
-          limit: DEFAULT_LIMIT,
         },
         method: "GET",
         result: {
@@ -246,37 +185,14 @@ const executionTests: ExecutionTestTableItem<Validators, OutputSchema>[] = [
     },
   },
   {
-    name: "user-provided settings override report settings",
+    name: "user-provided settings are sent to the API",
     apiCallHandler: requestsInOrder([
-      {
-        endpoint: `/v2/cost_reports/${pathEncode("crt_123")}` as const,
-        params: {},
-        method: "GET" as const,
-        result: {
-          ok: true as const,
-          data: {
-            ...reportData,
-            settings: {
-              ...REPORT_SETTINGS,
-              amortize: true,
-              include_credits: false,
-            },
-          },
-        },
-      },
       {
         endpoint: "/v2/costs",
         params: {
-          ...REPORT_SETTINGS_AS_API_PARAMS,
+          ...baseApiParams,
           "settings[amortize]": false,
           "settings[include_credits]": true,
-          cost_report_token: "crt_123",
-          start_date: "2023-01-01",
-          end_date: "2023-01-31",
-          date_bin: "month",
-          groupings: DEFAULT_GROUPINGS,
-          page: 1,
-          limit: DEFAULT_LIMIT,
         },
         method: "GET",
         result: {
@@ -295,94 +211,15 @@ const executionTests: ExecutionTestTableItem<Validators, OutputSchema>[] = [
     },
   },
   {
-    name: "report with amortize=false uses report default",
+    name: "unsuccessful call",
     apiCallHandler: requestsInOrder([
-      {
-        endpoint: `/v2/cost_reports/${pathEncode("crt_456")}` as const,
-        params: {},
-        method: "GET" as const,
-        result: {
-          ok: true as const,
-          data: {
-            ...reportData,
-            token: "crt_456",
-            settings: {
-              ...REPORT_SETTINGS,
-              amortize: false,
-            },
-          },
-        },
-      },
       {
         endpoint: "/v2/costs",
         params: {
-          ...REPORT_SETTINGS_AS_API_PARAMS,
-          "settings[amortize]": false,
-          cost_report_token: "crt_456",
-          start_date: "2023-01-01",
-          end_date: "2023-01-31",
-          date_bin: "month",
-          groupings: DEFAULT_GROUPINGS,
-          page: 1,
-          limit: DEFAULT_LIMIT,
-        },
-        method: "GET",
-        result: {
-          ok: true,
-          data: successData,
-        },
-      },
-    ]),
-    handler: async ({ callExpectingSuccess }) => {
-      const res = await callExpectingSuccess({
-        ...validArguments,
-        cost_report_token: "crt_456",
-      });
-      expect(res.costs).toEqual(successData.costs);
-    },
-  },
-  {
-    name: "unsuccessful cost report fetch",
-    apiCallHandler: requestsInOrder([
-      {
-        endpoint: `/v2/cost_reports/${pathEncode("crt_123")}` as const,
-        params: {},
-        method: "GET" as const,
-        result: {
-          ok: false as const,
-          errors: [{ message: "Cost report not found" }],
-        },
-      },
-    ]),
-    handler: async ({ callExpectingMCPUserError }) => {
-      const err = await callExpectingMCPUserError({
-        ...validArguments,
-        page: 1,
-        cost_report_token: "crt_123",
-        start_date: undefined,
-        end_date: undefined,
-        date_bin: undefined,
-      });
-      expect(err.exception).toEqual({
-        errors: [{ message: "Cost report not found" }],
-      });
-    },
-  },
-  {
-    name: "unsuccessful costs call",
-    apiCallHandler: requestsInOrder([
-      getReportCall("crt_123"),
-      {
-        endpoint: "/v2/costs",
-        params: {
-          cost_report_token: "crt_123",
-          page: 1,
+          ...baseApiParams,
           start_date: undefined,
           end_date: undefined,
           date_bin: undefined,
-          limit: DEFAULT_LIMIT,
-          groupings: DEFAULT_GROUPINGS,
-          ...REPORT_SETTINGS_AS_API_PARAMS,
         },
         method: "GET",
         result: {
