--- src/tools/users/index.ts
@@ -1,2 +1,0 @@
-import "./get-user";
-import "./get-users";
